requirements.txt
streamlit
google-generativeai
Pillow
matplotlib